crawler/requirements.txt
# VN-Law-Mini Crawler Dependencies
# Python 3.8+

# HTTP requests
requests==2.31.0

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Supabase client
supabase==2.3.0

# Environment variables
python-dotenv==1.0.0
